o replace IP of Nexus
Now that the NGT has a build that uses a local Nexus if there is one, it
is much easier to set a certain machine as the artifact repo. Actually
this step strictly is not necessary any more – one still needs to change
things by hand. However, not using "localhost" avoids the situation that the build
works on the integration machine, but for nobody else.